Migrate Glances to new entity naming style
