+++ homeassistant/components/glances/__init__.py
@@ -129,7 +129,7 @@ class GlancesData:
 def get_api(hass, entry):
     """Return the api from glances_api."""
     params = entry.copy()
-    params.pop(CONF_NAME)
+    params.pop(CONF_NAME, None)
     verify_ssl = params.pop(CONF_VERIFY_SSL, True)
     httpx_client = get_async_client(hass, verify_ssl=verify_ssl)
     return Glances(httpx_client=httpx_client, **params)
