+++ homeassistant/components/glances/sensor.py
@@ -1,7 +1,25 @@
 """Support gathering system information of hosts which are running glances."""
-from homeassistant.components.sensor import SensorEntity
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from homeassistant.components.sensor import (
+    SensorDeviceClass,
+    SensorEntity,
+    SensorEntityDescription,
+    SensorStateClass,
+)
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_NAME, STATE_UNAVAILABLE, Platform
+from homeassistant.const import (
+    CONF_HOST,
+    CONF_NAME,
+    DATA_GIBIBYTES,
+    DATA_MEBIBYTES,
+    PERCENTAGE,
+    STATE_UNAVAILABLE,
+    TEMP_CELSIUS,
+    Platform,
+)
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import entity_registry
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
@@ -9,7 +27,203 @@ from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import GlancesData
-from .const import DATA_UPDATED, DOMAIN, SENSOR_TYPES, GlancesSensorEntityDescription
+from .const import CPU_ICON, DATA_UPDATED, DOMAIN
+
+
+@dataclass
+class GlancesSensorEntityDescription(SensorEntityDescription):
+    """Describe Glances sensor entity."""
+
+    type: str | None = None
+    name_suffix: str | None = None
+
+
+SENSOR_TYPES: tuple[GlancesSensorEntityDescription, ...] = (
+    GlancesSensorEntityDescription(
+        key="disk_use_percent",
+        type="fs",
+        name_suffix="used percent",
+        native_unit_of_measurement=PERCENTAGE,
+        icon="mdi:harddisk",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="disk_use",
+        type="fs",
+        name_suffix="used",
+        native_unit_of_measurement=DATA_GIBIBYTES,
+        icon="mdi:harddisk",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="disk_free",
+        type="fs",
+        name_suffix="free",
+        native_unit_of_measurement=DATA_GIBIBYTES,
+        icon="mdi:harddisk",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="memory_use_percent",
+        type="mem",
+        name_suffix="RAM used percent",
+        native_unit_of_measurement=PERCENTAGE,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="memory_use",
+        type="mem",
+        name_suffix="RAM used",
+        native_unit_of_measurement=DATA_MEBIBYTES,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="memory_free",
+        type="mem",
+        name_suffix="RAM free",
+        native_unit_of_measurement=DATA_MEBIBYTES,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="swap_use_percent",
+        type="memswap",
+        name_suffix="Swap used percent",
+        native_unit_of_measurement=PERCENTAGE,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="swap_use",
+        type="memswap",
+        name_suffix="Swap used",
+        native_unit_of_measurement=DATA_GIBIBYTES,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="swap_free",
+        type="memswap",
+        name_suffix="Swap free",
+        native_unit_of_measurement=DATA_GIBIBYTES,
+        icon="mdi:memory",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="processor_load",
+        type="load",
+        name_suffix="CPU load",
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="process_running",
+        type="processcount",
+        name_suffix="Running",
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="process_total",
+        type="processcount",
+        name_suffix="Total",
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="process_thread",
+        type="processcount",
+        name_suffix="Thread",
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="process_sleeping",
+        type="processcount",
+        name_suffix="Sleeping",
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="cpu_use_percent",
+        type="cpu",
+        name_suffix="CPU used",
+        native_unit_of_measurement=PERCENTAGE,
+        icon=CPU_ICON,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="temperature_core",
+        type="sensors",
+        name_suffix="Temperature",
+        native_unit_of_measurement=TEMP_CELSIUS,
+        device_class=SensorDeviceClass.TEMPERATURE,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="temperature_hdd",
+        type="sensors",
+        name_suffix="Temperature",
+        native_unit_of_measurement=TEMP_CELSIUS,
+        device_class=SensorDeviceClass.TEMPERATURE,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="fan_speed",
+        type="sensors",
+        name_suffix="Fan speed",
+        native_unit_of_measurement="RPM",
+        icon="mdi:fan",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="battery",
+        type="sensors",
+        name_suffix="Charge",
+        native_unit_of_measurement=PERCENTAGE,
+        icon="mdi:battery",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="docker_active",
+        type="docker",
+        name_suffix="Containers active",
+        icon="mdi:docker",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="docker_cpu_use",
+        type="docker",
+        name_suffix="Containers CPU used",
+        native_unit_of_measurement=PERCENTAGE,
+        icon="mdi:docker",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="docker_memory_use",
+        type="docker",
+        name_suffix="Containers RAM used",
+        native_unit_of_measurement=DATA_MEBIBYTES,
+        icon="mdi:docker",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="used",
+        type="raid",
+        name_suffix="Raid used",
+        icon="mdi:harddisk",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    GlancesSensorEntityDescription(
+        key="available",
+        type="raid",
+        name_suffix="Raid available",
+        icon="mdi:harddisk",
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+)
 
 
 async def async_setup_entry(
@@ -20,7 +234,7 @@ async def async_setup_entry(
     """Set up the Glances sensors."""
 
     client: GlancesData = hass.data[DOMAIN][config_entry.entry_id]
-    name = config_entry.data[CONF_NAME]
+    name = config_entry.data.get(CONF_NAME, "")
     dev = []
 
     @callback
@@ -102,6 +316,8 @@ class GlancesSensor(SensorEntity):
     """Implementation of a Glances sensor."""
 
     entity_description: GlancesSensorEntityDescription
+    _attr_has_entity_name = True
+    _attr_should_poll = False
 
     def __init__(
         self,
@@ -117,11 +333,15 @@ class GlancesSensor(SensorEntity):
         self.unsub_update = None
 
         self.entity_description = description
-        self._attr_name = f"{name} {sensor_name_prefix} {description.name_suffix}"
+        self._attr_name = f"{sensor_name_prefix} {description.name_suffix}"
+        # if name exists preserve previous naming
+        if name:
+            self._attr_has_entity_name = False
+            self._attr_name = f"{name} {self._attr_name}"
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, glances_data.config_entry.entry_id)},
             manufacturer="Glances",
-            name=name,
+            name=name or glances_data.config_entry.data[CONF_HOST],
         )
         self._attr_unique_id = f"{self.glances_data.config_entry.entry_id}-{sensor_name_prefix}-{description.key}"
 
@@ -134,11 +354,6 @@ class GlancesSensor(SensorEntity):
     def native_value(self):
         """Return the state of the resources."""
         return self._state
-
-    @property
-    def should_poll(self):
-        """Return the polling requirement for this sensor."""
-        return False
 
     async def async_added_to_hass(self):
         """Handle entity which will be added."""
