+++ homeassistant/components/glances/strings.json
@@ -2,9 +2,7 @@
   "config": {
     "step": {
       "user": {
-        "title": "Setup Glances",
         "data": {
-          "name": "[%key:common::config_flow::data::name%]",
           "host": "[%key:common::config_flow::data::host%]",
           "username": "[%key:common::config_flow::data::username%]",
           "password": "[%key:common::config_flow::data::password%]",
