+++ homeassistant/components/glances/translations/en.json
@@ -11,15 +11,13 @@
             "user": {
                 "data": {
                     "host": "Host",
-                    "name": "Name",
                     "password": "Password",
                     "port": "Port",
                     "ssl": "Uses an SSL certificate",
                     "username": "Username",
                     "verify_ssl": "Verify SSL certificate",
                     "version": "Glances API Version (2 or 3)"
-                },
-                "title": "Setup Glances"
+                }
             }
         }
     },
